course ownership

+++ controllers/courses.js
@@ -49,12 +49,18 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @rout     POST /api/v1/bootcamps/:bootcampid/courses
 // @access   Public
 exports.addCourse = asyncHandler(async(req, res, next) => {
+
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404);
+    }
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to ${bootcamp._id}`), 401);
     }
 
     const course = await Course.create(req.body);
@@ -74,6 +80,10 @@ exports.updateCourese = asyncHandler(async(req, res, next) => {
 
     if(!course){
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`));
+    }
+
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`), 401);
     }
 
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -98,6 +108,11 @@ exports.deleteCourse = asyncHandler(async(req, res, next) => {
     if(!course){
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404);
     }
+
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`), 401);
+    }
+
 
     await course.deleteOne();
